+++ src/actions.ts
@@ -3,7 +3,7 @@ export {editorDevToolsToggle, editorOpen, editorInsert} from './modules/editor/a
 export {editorSave, editorSet, editorWriteFileFromFile, editorWriteFileFromContent} from './modules/editor/actions';
 export {hintPositionSet} from './modules/hints/actions';
 export {pageSet, pageNext} from './modules/page/actions';
-export {progressLoad, progressCompletePage} from './modules/progress/actions';
+export {progressCompletePage, progressLoad, progressReset} from './modules/progress/actions';
 export {routeSet} from './modules/route/actions';
 export {setupVerify, setupPackage} from './modules/setup/actions';
 export {testLoad, testRun, testResult, testComplete} from './modules/tests/actions';